Split the long servlett class
